# Use math instead of numpy for calculating TTC, on average 80% faster