Code I used to benchmark: https://pastebin.com/ryBV3Lby

Benchmark:

```
run(1000)
Numpy took 0.005 seconds
Math took 0.001 seconds
79.9933% faster
run(10000)
Numpy took 0.04801 seconds
Math took 0.008 seconds
83.3333% faster
run(100000)
Numpy took 0.49011 seconds
Math took 0.08602 seconds
82.4489% faster
run(1000000)
Numpy took 4.97812 seconds
Math took 0.84819 seconds
82.9617% faster
```